refactor: remove duplicate PipelinePolicy tests and move interaction tests to subdirectory
